This version is as far as I can tell working.
2:00pm - I couldn't figure how to upload to dropbox.
This version is as far as I can tell working.  I tried to simplify
and shorten a lot of the code.  You no longer have to press clear to
index and retrieve multiple requirements.  The Set Overlap works
decently.  I do not think it would be too difficult to add other
methods before the meeting, if someone has time/wants to.